update: move convert() to other folder
